enforce required argument

+++ lib/delighted/resources/survey_request.rb
@@ -3,6 +3,7 @@ module Delighted
     self.path = "/people/{PERSON_EMAIL}/survey_requests"
 
     def self.delete_pending(attributes = {}, client = Delighted.shared_client)
+      raise ArgumentError, "You must pass :person_email" unless attributes[:person_email]
       interpolated_path = path.sub("{PERSON_EMAIL}", CGI.escape(attributes[:person_email]))
       interpolated_path << "/pending"
       json = client.delete_json(interpolated_path)
